PoC with UX realized in browser console
Detects the following fields:
* URL
* title
* author (poor coverage)
* publish date
* access date
* publisher
* language
* archive-url and archive-date
* quote

Archive.org support is automatic. When the extension is used on archived
page loaded from web.archive.org it will populate: original URL field,
archived URL and archive date.

If text has been selected on the page, when extension was triggered,
the selection is included as a quote.

Special chars breaking Mediawiki markup (ie. |, [, ], }, {) are
converted to entities.